refactor(frontend/admin): logout user (#127)
